requirements.txt
azure-common==1.1.25
azure-core==1.7.0
azure-storage-common==2.1.0
azure-storage-file==2.1.0
azure-storage-file-share==12.1.2
astroid==2.3.1
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.0
cftime==1.0.3.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
DateTime==4.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
motuclient==1.8.4
netCDF4==1.5.2
numpy==1.17.2
pandas==0.25.1
pycparser==2.19
pylint==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
scipy==1.3.1
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
WTForms==2.2.1
zope.interface==4.6.0
